move and update all to .NET 5.0
