+++ src/cloud/IoTalkie.Common/AzureSettings.cs
