+++ src/cloud/IoTalkie.Common/UpdateFirmwareCmd.cs
@@ -1,0 +1,18 @@
+﻿using System;
+
+namespace IoTalkie.MessageSenderCmd
+{
+    public class UpdateFirmwareCmd
+    {
+        public string Cmd { get; set; } = "updateFirmware";
+
+        public string MessageId { get; set; } = Guid.NewGuid().ToString();
+
+        public long Timestamp { get; set; } = DateTimeOffset.Now.ToUnixTimeSeconds();
+
+
+        public long Size { get; set; }
+
+        public string RemoteUrl { get; set; }
+    }
+}
