+++ src/cloud/IoTalkie.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 16.0.30907.101
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IoTalkie", "IoTalkie\IoTalkie.csproj", "{F7926BA6-B848-4C48-B1DA-DE2259F488A3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IoTalkie.MessageSenderCmd", "IoTalkie.MessageSenderCmd\IoTalkie.MessageSenderCmd.csproj", "{835D0510-3CC2-46F8-881D-D8BD5A749D7F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IoTalkie.MessageSenderCmd", "IoTalkie.MessageSenderCmd\IoTalkie.MessageSenderCmd.csproj", "{835D0510-3CC2-46F8-881D-D8BD5A749D7F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IoTalkie.Common", "IoTalkie.Common\IoTalkie.Common.csproj", "{4E3EBB37-F1D5-49E9-A3EC-552702F8E7DF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{835D0510-3CC2-46F8-881D-D8BD5A749D7F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{835D0510-3CC2-46F8-881D-D8BD5A749D7F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{835D0510-3CC2-46F8-881D-D8BD5A749D7F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4E3EBB37-F1D5-49E9-A3EC-552702F8E7DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4E3EBB37-F1D5-49E9-A3EC-552702F8E7DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4E3EBB37-F1D5-49E9-A3EC-552702F8E7DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4E3EBB37-F1D5-49E9-A3EC-552702F8E7DF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
